+++ mobile/components/ui/HapticCheckbox.js
@@ -1,13 +1,6 @@
-import React from 'react';
 import { Checkbox } from 'react-native-paper';
-import * as Haptics from 'expo-haptics';
+import { withHapticFeedback } from './hapticUtils';
 
-const HapticCheckboxItem = ({ onPress, ...props }) => {
-  const handlePress = (e) => {
-    Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-    if (onPress) onPress(e);
-  };
-  return <Checkbox.Item onPress={handlePress} {...props} />;
-};
+const HapticCheckboxItem = withHapticFeedback(Checkbox.Item);
 
 export default HapticCheckboxItem;